trinity/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
ez-setup==0.9
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysqlclient==1.3.14
protobuf==3.6.1
PyMySQL==0.9.3
pytz==2018.9
six==1.12.0
